***Replaces all files.***
Index.html remains mostly unchanged.
Feedback.html is a barebones template with an empty form on it.
Main-1.scss is the main css file - A much cleaner and stable version of the previous main.scss.
Form-validation-kit is added in file structure and linked on feedback.html but not implemented yet.

Working on iPhone 4/ iPad 3/4 and Desktop.